Rename TypeScript object utility to match canonial declaration
