+++ packages/voictents-and-estinants-engine/src/adapted-programs/programs/render-knowledge-graph/constructDynamicMetadataFile.ts
@@ -8,7 +8,7 @@ import {
 import {
   TypeScriptArray,
   TypeScriptObjectInstance,
-} from '../../../package-agnostic-utilities/object/object';
+} from '../../../package-agnostic-utilities/object/typeScriptObject';
 import {
   OUTPUT_FILE_GEPP,
   OutputFile,
