+++ packages/voictents-and-estinants-engine/src/package-agnostic-utilities/constructor-function/types.ts
@@ -3,7 +3,7 @@
  */
 
 import { UnsafeTuple } from '../type/tuple';
-import { TypeScriptObjectInstance } from '../object/object';
+import { TypeScriptObjectInstance } from '../object/typeScriptObject';
 
 export type ConstructorFunctionName = string;
