+++ packages/voictents-and-estinants-engine/src/package-agnostic-utilities/object/typeScriptObject.ts
