+++ packages/voictents-and-estinants-engine/src/package-agnostic-utilities/type/combination.ts
@@ -1,4 +1,4 @@
-import { TypeScriptObjectInstanceKey } from '../object/object';
+import { TypeScriptObjectInstanceKey } from '../object/typeScriptObject';
 
 export type Combination<TValue extends TypeScriptObjectInstanceKey> = Record<
   TValue,
