+++ packages/voictents-and-estinants-engine/src/package-agnostic-utilities/typed-datum/customTypedDatum.ts
@@ -4,7 +4,7 @@ import {
   TypeScriptMap,
   TypeScriptObjectInstance,
   TypeScriptSet,
-} from '../object/object';
+} from '../object/typeScriptObject';
 import {
   getTypeScriptTypedDatum,
   TypeScriptDatumTypeName,
