+++ packages/voictents-and-estinants-engine/src/package-agnostic-utilities/typed-datum/typeScriptTypedDatum.ts
@@ -1,5 +1,5 @@
 import { TypeScriptFunction } from '../function/typeScriptFunction';
-import { TypeScriptObject } from '../object/object';
+import { TypeScriptObject } from '../object/typeScriptObject';
 
 export enum TypeScriptDatumTypeName {
   bigint = 'bigint',
